lexer: reject ]]> sequence in CharData production
The `]]>` sequence is forbidden as per the XML standard [1]. This
change advised some refactoring because now a lot of code was
suddenly shared between the (now removed) Content::CDataSection
state and the Content::Initial state (which would have had to
handle the same exit sequence).

Thanks to moparisthebest for an extensive AFL-generated test
corpus.

   [1]: XML 1.0 § 2.4 [14] CharData: https://www.w3.org/TR/REC-xml/#NT-CharData