lexer: Do not keep scratchpad memory allocated
Previously, we used split_off() to create a new perfectly fitting
string from the data in the scratchpad, leaving the scratchpad
allocation alive. The idea behind that was that this avoids
regrowing the scratchpad for each token again, which costs CPU
time and may cause memory fragmentation.

However, keeping the scratchpad allocated is quite unfortunate
when dealing with many long-lived XML parsers which are idle most
of the time. Then, the scratchpad memory grows up to (potentially)
max token size and is not put to any good use.

Ideally, there would be a way for the application to signal that
such temporary memory can "now" be dropped because a chunk of data
has been processed. We don’t have that yet :).