Allow releasing temporary buffers explicitly
The motivation for this is in a situation where the application
knows (or can assume reasonably) that it will be a while until the
parser is next needed. In such cases, the cost of the memory which
is unusable for other parts of the application (e.g. other
parsers) may be higher than the CPU and time cost of reallocating
the buffers when the parser is needed next.