lexer: avoid unnecessary allocation
The idea here is that, hypothetically, if Name or CData had a
string type which was able to use some of the 24 bytes needed for
a String to inline the data if the string is short enough, we
avoid an allocation of a String by not using split_off but instead
constructing the object from the string reference to the
scratchpad.

This is a foreshadowing of the migration to smartstring for Name,
NCName and potentially CData.