strings: move to smartstring and consolidate things
Sorry, this commit should’ve been split, but doing so would be a
major PITA because I forgot to commit after the smartstring stuff.

Either way, this consolidates the traits implemented by the
different string and str types exposed by the library to make them
more uniform to use.

In addition, Name and NCName are ported to use smartstring
internally. smartstring embeds strings shorter than 24 bytes in the
space required for the String struct instead of allocating the
string on the heap. For many XML workloads, this can save a
significant amount of allocations with very low runtime overhead
and zero memory overhead.