Increase parser performance by about 40%
This refactor/rewrite is based on two observations. First, the utf8dec
and q benchmarks show that the io::Read implementation of &[u8] is
blazingly fast. Yet, BufferedQueue::read() is still very very expensive
compared to that.

That means we have to use a slice of bytes in lexing where ever we can.

The second observation is that Utf8Decoder is much slower than the
standard library one: we're talking a factor of 10. So we now just
don't do that anymore.

Of course, we still have to decode UTF-8 on input documents, but we
delay that until we actually emit strings from the lexer: all
delimiters in XML are ASCII, and ASCII is a strict subset of UTF-8. That
in turn means that all lexing can take place on u8 instead of the
Utf8Char structs we used before.

Some special care had to be taken around splitting UTF-8 sequences
across read buffer boundaries, but that works just fine (there are
tests to prove it ☺).