lexer: fix handling of eof
The inline comments should already explain a lot of what is going on
here, but to summarize, this establishes compatibility with
edge-triggered and very simple read loops.

The previous setup had the issue that the Lexer was never sure if the
end of a buffer meant that Eof had been reached or whether it just meant
that the source has currently no more data available.

The new loop in the lex function resolves this ambiguity: if the lexer
generates a WouldBlock despite a previous successful read, it will
simply try to read again until either the source returns an error or
the lexer emits a token or a non-WouldBlock error.

This means that lex() may now issue multiple calls to the underlying
data source, but this is ok: Either the source is blocking, in which
case even the first call to `fill_buf` may cause blocking I/O, or the
source is non-blocking, in which case the call will fail quickly with
WouldBlock. It essentially takes the load off the caller to loop this.