Re-write async support more idiomatically
This allows us to get rid of async-trait. async-trait has the downside
that the trait functions all return some Box<dyn Future>, which prevents
the compiler from seeing useful types.

The new implementation uses explicit structs for the high-level
operations and a trait with only polling functions for the low-level
asynchronicity, just like tokio does with AsyncRead and AsyncReadExt for
instance.

It just took a few productive weeks of async rust for me to be able to
come back to this and rewrite it this way.