Refactor strings module into rxml_validation
This resolves a circular (dev-)dependency between rxml_proc and
rxml. The circular requirement existed because rxml_proc needed the
string definitions in its dev-dependencies. By moving those into
rxml_validation, we avoid this.

This comes at the cost of increasing the footprint of the proc macro
crate dependencies, but I guess we cannot really avoid that for now.