Redefine PushDriver as a thing taking bytes::Buf
Previously, the PushDriver would focus inherently around the
BufferQueue. The BufferQueue is a design from the olden days. It was
intended to allow zero-copy use of arbitrary buffers during streamed
parsing. However, that can only be safely done by either wrapping them
in some Arc-y thing (which doesn't work because of the Cow type bound)
or if a lifetime is available which is bounded by the lifetime of the
entire parsed document.

The problem with the latter is that then you could also just use the
PullDriver, and it'd probably be more efficient, too.

The solution to this is to use bytes::Buf. Bufs can be read partially,
which means that the driver does not need to keep a reference to the Buf
internally; we just make it the task of the caller to repeatedly provide
the same Buf when reading from the same document, until that Buf is
depleted.

The only caveat is that the eof-ness of a buffer is now to be explicitly
communicated via an argument.

Oh and also this completely breaks the PushDriver API. But that is what
0.x versions are for.