Fold NWFError and WFError into a single XmlError type
Especially typing out the `Error:Not{Namespace,}WellFormed` variants was
a major pain when working with the error types. In addition, the line
between namespace- and normal well-formedness was already blurry: for
instance, WFError::DuplicateAttribute was *also* reported for
a namespace-well-formed constraint because of efficiency.

Even though breaking, this should simplify the API significantly.