Apply rust naming guidelines around string types
See-Also: https://rust-lang.github.io/api-guidelines/naming.html

`CData` and `NCName` were particularly hard calls. The document linked
above gives `Usize` as an example of a correct name, as well as `Stdin`.
Applying that principle under the assumption that `Usize` stems from
"unsigned size" and `Stdin` from "standard input", the application of
those rules to `CData` ("character data") and `NCName` (non-colon name)
would result in the ugly and confusing `Cdata` and `Ncname` type names.

However, if we assume that the page above uses `usize` and `stdin` as
atomic tokens and the capitalization rules stem from that, we can make
up a rule "words get capitalized and acronyms are words on their own".
Based on that, `CData` and `NcName` are derived.

I can live with that.