writer: Fix use of explicitly prefixed elements
This fixes the emission of prefixed elements. SimpleNamespaces, by
default, will never cause prefixed elements. However, by declaring a
prefix before handing the SimpleNamespaces instance to the Encoder, by
using a custom TrackNamespace implementation or by future exposure of
the TrackNamespace implementation from the Encoder, it is possible to
trigger this situation.

Previously, the code would omit the localname, which is arguably
terrible.